+++ core/src/com/aneebo/rotg/abilities/defense/Teleport.java
@@ -21,8 +21,8 @@ public class Teleport extends Ability {
 	
 	private Array<Entity> targets;
 	
-	public Teleport(int id, int castTime, float range, AbilityType type,
-			String name, int cooldown, float damage, float energy_cost,
+	public Teleport(int id, float castTime, float range, AbilityType type,
+			String name, float cooldown, float damage, float energy_cost,
 			String textureName, Engine engine, Array<Upgrade> upgrades) {
 		super(id, castTime, range, type, name, cooldown, damage, energy_cost,
 				textureName, engine, upgrades);
